// src/shared/api/generated/model/region.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * 아이랑 - 홈 화면 API
 * OpenAPI spec version: 0.0.1
 */

export type Region = typeof Region[keyof typeof Region];


export const Region = {
  SEOUL: 'SEOUL',
  BUSAN: 'BUSAN',
  DAEGU: 'DAEGU',
  INCHEON: 'INCHEON',
  GWANGJU: 'GWANGJU',
  DAEJEON: 'DAEJEON',
  ULSAN: 'ULSAN',
  SEJONG: 'SEJONG',
  GYEONGGI: 'GYEONGGI',
  GANGWON: 'GANGWON',
  CHUNGBUK: 'CHUNGBUK',
  CHUNGNAM: 'CHUNGNAM',
  JEONBUK: 'JEONBUK',
  JEONNAM: 'JEONNAM',
  GYEONGBUK: 'GYEONGBUK',
  GYEONGNAM: 'GYEONGNAM',
  JEJU: 'JEJU',
} as const;